Roles for user and organization module - Added roles for user module with proper endpoints - Moved logic from role module to organization module responsible for assinging, deleting and getting roles per user in organization
