+++ app/users/user_manager.py
@@ -2,6 +2,7 @@ from typing import Optional
 
 from fastapi import Request
 
+from app.roles.schemas import RoleFields, UserRolesFields
 from app.users.schemas import SearchableUserFields, CreateUserFields, UpdateUserFields, UserFields
 from app.users.users_manager_api_layer import UserManagerApiLayer
 from app.config import Settings
@@ -64,6 +65,44 @@ class UserManager:
             content=user_fields.model_dump_json(exclude_none=True)
         )
         return UserFields(**created_user_data)
+
+    async def assign_user_roles(
+            self,
+            auth_token: str,
+            user_id: str,
+            members_roles_fields: UserRolesFields
+    ) -> None:
+        await self._api_layer.make_request(
+            method="POST",
+            endpoint=f'/users/{user_id}/roles',
+            auth_token=auth_token,
+            content=members_roles_fields.model_dump_json(exclude_none=True)
+        )
+
+    async def delete_user_roles(
+            self,
+            auth_token: str,
+            user_id: str,
+            members_roles_fields: UserRolesFields
+    ) -> None:
+        await self._api_layer.make_request(
+            method="DELETE",
+            endpoint=f'/users/{user_id}/roles',
+            auth_token=auth_token,
+            content=members_roles_fields.model_dump_json(exclude_none=True)
+        )
+
+    async def get_user_roles(
+            self,
+            auth_token: str,
+            user_id: str
+    ) -> list[RoleFields] | list:
+        user_roles = await self._api_layer.make_request(
+            method="GET",
+            endpoint=f'/users/{user_id}/roles',
+            auth_token=auth_token,
+        )
+        return [RoleFields(**user_role) for user_role in user_roles]
 
 
 def get_user_manager_service(request: Request) -> UserManager:
